Integrate Decap CMS
