+++ content/publicacoes/exemplo.md
@@ -1,0 +1,11 @@
+---
+title: "Boletim Econômico do Rio – 1ª edição – 2025"
+category: "BOLETIM ECONÔMICO"
+year: 2025
+autor: "Equipe"
+resumo: "Resumo da publicação"
+conteudo: "Texto completo"
+file: "boletim-economico-do-rio-1a-edicao-2025.pdf"
+---
+
+Conteúdo opcional em Markdown.
