+++ package.json
@@ -19,7 +19,8 @@
         "react-icons": "^5.4.0",
         "recharts": "^2.15.0",
         "tailwind-merge": "^2.6.0",
-        "typescript": "^5.7.2"
+        "typescript": "^5.7.2",
+        "gray-matter": "^4.0.3"
     },
     "devDependencies": {
         "@tailwindcss/typography": "^0.5.15",
