+++ public/admin/index.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="UTF-8" />
+    <title>Decap CMS</title>
+    <script src="https://unpkg.com/decap-cms@^2/dist/decap-cms.js"></script>
+  </head>
+  <body>
+    <!-- Decap CMS will inject the interface here -->
+  </body>
+</html>
